Huge update
main.cpp:
- moved all functions to its own files
- fixed '-Wwrite-strings' on fileToBytes()
- fixed byte offset for shifts in files
- fixed length of files
- added custom throw (WIP) (see src/includes/exceptions.h)
- main.cpp now only has main() loop

build.sh
- fixed includes

io.h
- fixed fileToBytes arguments
