+++ src/includes/math.h
@@ -1,0 +1,7 @@
+struct bitsInByte{		// data object for bits, stores bits generated by byteToBit()
+	char bits[8];
+};
+
+bitsInByte byteToBit(char byteIn); // makes bits out of bytes, welp
+char* unsignedlonglongToFormattedHex(unsigned long long longlongIn); // formats bytesOutputWithData.size to output address section (in hex)
+
